requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
pydantic[email]==2.3.0
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.24.1
python-dotenv==1.0.0
cryptography==41.0.3
bcrypt==4.0.1
alembic==1.12.0
PyPDF2==3.0.1 
EbookLib==0.18.0
gunicorn==21.2.0

pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
sqlalchemy-utils==0.41.1
faker==19.6.2